The first argument lat_aux_grid, should be numeric and not integer.
This function has not been tested by me because I don't have any
tests for it yet.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11140 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/lib/nfp/wrapper.c
@@ -1751,7 +1751,7 @@ void NclAddUserFuncs(void)
     nargs = 0;
     args = NewArgs(6);
     
-    SetArgTemplate(args,nargs,"integer",1,NclANY);nargs++;
+    SetArgTemplate(args,nargs,"numeric",1,NclANY);nargs++;
     SetArgTemplate(args,nargs,"numeric",3,NclANY);nargs++;
     SetArgTemplate(args,nargs,"numeric",3,NclANY);nargs++;
     SetArgTemplate(args,nargs,"numeric",3,NclANY);nargs++;
